Too bad can't solve it in C++
I write the code in C++ however it doesn't have the Biginterger class
like Java and I feel not well to use the gmp library. So I rewrite the
code in python and it's really fast under pypy.  I still wanna try some
way in C++ to solve this. And I begin to think why python plays an
important role in big Data analysis. So easy and so quick. C++ is
better for algorithm and data structure. C is the best when it comes to
the operation system and especially when you have to work with the
lower part of the machine. That’s a big difference.